Updated from home repository
git-svn-id: https://svn.aros.org/svn/aros/trunk@34109 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/codesetslib/developer/examples/b64d.c
@@ -58,7 +58,7 @@ int main(int argc,char **argv)
             r = CodesetsDecodeB64(CSA_B64SourceFile, (Tag)argv[1],
                                   CSA_B64DestFile,   (Tag)argv[2],
                                   TAG_DONE);
-            printf("Res %d\n",r);
+            printf("Res %d\n", (int)r);
 
             DROPINTERFACE(ICodesets);
             CloseLibrary(CodesetsBase);
